% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSOM.R
\name{getClusterData}
\alias{getClusterData}
\title{Retrieve Clustered Data}
\usage{
getClusterData()
}
\value{
A data frame with the clustered dataset.
}
\description{
Access the dataset with cluster assignments stored by `clusterSOM`.
}
