% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSOM.R
\name{clusterSOM}
\alias{clusterSOM}
\title{Perform Clustering on SOM Nodes}
\usage{
clusterSOM(model, plot_result = TRUE, file_path = NULL)
}
\arguments{
\item{model}{The trained SOM model object.}

\item{plot_result}{A logical value indicating whether to plot the clustering result. Default is `TRUE`.}

\item{file_path}{An optional string specifying the path to a CSV file. If provided, clusters are assigned to the observations in the original dataset, and the updated data with an added "Cluster" column is stored globally as 'Data&Clusters'.}
}
\value{
A plot of the clusters on the SOM grid (if `plot_result = TRUE`). Additionally, if `file_path` is specified, a modified dataset with cluster assignments is stored in the global environment.
}
\description{
Groups similar nodes of the SOM using hierarchical clustering and the KGS
penalty function to determine the optimal number of clusters.
}
\examples{
\dontrun{
  clusterSOM(model, plot_result = TRUE)
  clusterSOM(model, plot_result = FALSE, file_path = "data.csv")
}
}
