% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalSOM.R
\name{finalSOM}
\alias{finalSOM}
\title{Train Final SOM Model}
\usage{
finalSOM(data, dimension, iterations)
}
\arguments{
\item{data}{The preprocessed data matrix containing the input data for SOM training.}

\item{dimension}{An integer specifying the dimension of the square SOM grid (e.g., 5 results in a 5x5 grid).}

\item{iterations}{An integer defining the number of iterations for training the SOM model. Use a large value, e.g., 500 or higher, for improved training (an error message could suggest that reducing the number of iterations might be necessary).}
}
\value{
A trained SOM model object.
}
\description{
Re-trains the SOM using a specified optimal grid size and number of iterations.
}
\examples{
\dontrun{
  myFinalSOM <- finalSOM(data, dimension = 6, iterations = 700)
}
}
