% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePlot.R
\name{generatePlot}
\alias{generatePlot}
\title{Generate SOM Visualization Plots}
\usage{
generatePlot(model, plot_type, data = NULL)
}
\arguments{
\item{model}{The trained SOM model object.}

\item{plot_type}{An integer specifying the type of plot to generate. Options are:
\describe{
  \item{1}{Training progress plot (changes during training).}
  \item{2}{Node count plot (number of samples mapped to each node) for assessing map quality.}
  \item{3}{U-matrix plot (visualizing similarities between neighboring nodes).}
  \item{4}{Weight vector plot (patterns in the distributions of variables).}
  \item{5}{Kohonen heatmaps for all variables in the dataset (distribution of single variables across the map).}
}}

\item{data}{The preprocessed data matrix containing the input data. Required only for `plot_type = 5`.}
}
\value{
A plot or a series of plots is generated and displayed based on the specified type.
}
\description{
Creates various types of plots to visualize and evaluate the trained SOM model.
}
\examples{
\dontrun{
  generatePlot(model, plot_type = 2)
  generatePlot(model, plot_type = 5, data)
}
}
